# docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "swagger": "2.0",
    "info": {
        "description": "![Smart](http://localhost:8000/docs/file/logo.png)\r\n\r\n# 概述\r\n\r\n轻量易用的授权基座smart，提供一个非官方RBAC授权模型的基础开箱即用中台管理服务。\r\n\r\n支持响应码自定义国际化、高粒度权限分配、接口路由配置和实时生效，同时支持部门数据权限横向访问限制。\r\n\r\n本系统的全部接口采用【POST】【application/json】方式传输数据。\r\n\r\n除开放接口以外的其他接口均需要通过【ApiKeyAuth:请求头[Token]】完成鉴权。\r\n\r\n系统技术栈：Golang、VueNext、MySQL、Redis、Gin、ArcoDesign\r\n\r\n# 接口文档\r\n\r\n轻度依赖集成，通过静态HTML加载yaml文件进行打开接口文档。\r\n\r\n该接口文档提供Swagger[支持调试]和ReDoc[阅读增强]两个版本。\r\n\r\n[Swagger[支持调试]：/docs/swagger/index.html](http://localhost:8000/docs/swagger/index.html) 丨 [ReDoc[阅读增强]：/docs/redoc/index.html](http://localhost:8000/docs/redoc/index.html)\r\n\r\n## 生成说明\r\n\r\n- 首次集成\r\n\r\ngo install github.com/swaggo/swag/cmd/swag@latest\r\n\r\n- 格式化注释代码\r\n\r\nswag fmt\r\n\r\n- 初始化swagger.yaml文件\r\n\r\nswag init\r\n\r\n- 如果像本工程一样依赖了api.md文件描述项目\r\n\r\nswag init --md .\r\n\r\n\r\n- 删除多余的生成\r\n\r\nrm .\\docs\\docs.go\r\n\r\n删除docs.go是因为本项目中并未采用下述依赖来集成。\r\n\r\n而是通过HTML模板+JS+YAML引入集成，移除docs.go用于避免依赖编译报错。\r\n\r\n- github.com/swaggo/swag\r\n- github.com/swaggo/gin-swagger\r\n- github.com/swaggo/files\r\n\r\n源帮助页：[https://github.com/swaggo/swag/blob/master/README_zh-CN.md](https://github.com/swaggo/swag/blob/master/README_zh-CN.md)\r\n",
        "title": "smart",
        "contact": {
            "name": "smart",
            "url": "https://smart.siteol.com",
            "email": "dev@siteol.com"
        },
...
        }
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/plat/response/add": {
            "post": {
                "security": [
                    {
                        "Token": []
                    }
                ],
                "description": "新建响应码配置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "响应配置"
                ],
                "summary": "响应码新建",
                "operationId": "AddResponse 响应码新建",
                "parameters": [
                    {
                        "description": "请求",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/platModel.ResponseAddReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "响应成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/baseModel.ResBody"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/plat/response/del": {
            "post": {
                "security": [
                    {
                        "Token": []
                    }
                ],
                "description": "响应码封存处理",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "响应配置"
                ],
                "summary": "响应码封存",
                "operationId": "DelResponse 响应码封存",
                "parameters": [
                    {
                        "type": "string",
                        "description": "语言，不传默认为zh-CN",
                        "name": "Lang",
                        "in": "header"
                    },
                    {
                        "description": "请求",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/baseModel.IdReq"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "响应成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/baseModel.ResBody"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/plat/response/edit": {
            "post": {
                "security": [
                    {
                        "Token": []
                    }
                ],
                "description": "在响应码分组下编辑响应码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "响应配置"
                ],
                "summary": "响应码编辑",
                "operationId": "EditResponse 响应码编辑",
                "parameters": [
                    {
                        "description": "请求",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/platModel.ResponseEditReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "响应成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/baseModel.ResBody"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/plat/response/get": {
            "post": {
                "security": [
                    {
                        "Token": []
                    }
                ],
                "description": "查询响应码详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "响应配置"
                ],
                "summary": "响应码详情",
                "operationId": "GetResponse 响应码详情",
                "parameters": [
                    {
                        "description": "请求",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/baseModel.IdReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "响应成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/baseModel.ResBody"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/platModel.ResponseGetRes"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/plat/response/nextVal": {
            "post": {
                "security": [
                    {
                        "Token": []
                    }
                ],
                "description": "在业务分组下读取下一个响应码的建议值",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "响应配置"
                ],
                "summary": "响应码NextVal建议",
                "operationId": "NextResponseVal 响应码NextVal建议",
                "parameters": [
                    {
                        "description": "请求",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/platModel.ResponseNextValReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "响应成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/baseModel.ResBody"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "integer"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/plat/response/page": {
            "post": {
                "security": [
                    {
                        "Token": []
                    }
                ],
                "description": "查询响应码分页数据",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "响应配置"
                ],
                "summary": "响应码分页",
                "operationId": "PageResponse 响应码分页",
                "parameters": [
                    {
                        "description": "请求",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/platModel.ResponsePageReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "响应成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/baseModel.ResBody"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/baseModel.PageRes"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "list": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/platModel.ResponsePageRes"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "baseModel.IdReq": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "数据ID",
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "baseModel.PageRes": {
            "type": "object",
            "properties": {
                "list": {
                    "description": "分页数据"
                },
                "total": {
                    "description": "数据总量",
                    "type": "integer",
                    "example": 100
                }
            }
        },
        "baseModel.ResBody": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应码",
                    "type": "string",
                    "example": "S0000/F0000"
                },
                "data": {
                    "description": "响应数据"
                },
                "msg": {
                    "description": "响应消息",
                    "type": "string",
                    "example": "Success/Fail"
                },
                "unPop": {
                    "description": "免弹窗提示",
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "platModel.ResponseAddReq": {
            "type": "object",
            "required": [
                "enUs",
                "serviceCode",
                "type",
                "zhCn"
            ],
            "properties": {
                "code": {
                    "description": "响应码，仅示例，实际入库实时计算",
                    "type": "string",
                    "example": "S101"
                },
                "enUs": {
                    "description": "英文响应文言",
                    "type": "string",
                    "maxLength": 128,
                    "example": "EnUs"
                },
                "remark": {
                    "description": "其他备注信息",
                    "type": "string",
                    "maxLength": 64,
                    "example": "Remark"
                },
                "serviceCode": {
                    "description": "业务ID，来源于字典，指定响应码归属业务",
                    "type": "string",
                    "maxLength": 3,
                    "example": "0"
                },
                "type": {
                    "description": "响应类型，该字段用于筛选，可配置S/F/E",
                    "type": "string",
                    "enum": [
                        "S",
                        "F",
                        "E"
                    ],
                    "example": "S"
                },
                "zhCn": {
                    "description": "中文响应文言",
                    "type": "string",
                    "maxLength": 128,
                    "example": "ZhCn"
                }
            }
        },
        "platModel.ResponseEditReq": {
            "type": "object",
            "required": [
                "enUs",
                "id",
                "zhCn"
            ],
            "properties": {
                "enUs": {
                    "description": "英文响应文言",
                    "type": "string",
                    "maxLength": 128,
                    "example": "EnUs"
                },
                "id": {
                    "description": "数据ID",
                    "type": "integer",
                    "example": 1
                },
                "remark": {
                    "description": "其他备注信息",
                    "type": "string",
                    "maxLength": 64,
                    "example": "Remark"
                },
                "zhCn": {
                    "description": "中文响应文言",
                    "type": "string",
                    "maxLength": 128,
                    "example": "ZhCn"
                }
            }
        },
        "platModel.ResponseGetRes": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应码",
                    "type": "string",
                    "example": "20101"
                },
                "enUs": {
                    "description": "英文响应文言",
                    "type": "string",
                    "example": "EnUs"
                },
                "id": {
                    "description": "数据ID",
                    "type": "integer",
                    "example": 1
                },
                "remark": {
                    "description": "其他备注信息",
                    "type": "string",
                    "example": "Remark"
                },
                "serviceCode": {
                    "description": "业务ID，来源于字典，指定响应码归属业务",
                    "type": "string",
                    "example": "1"
                },
                "type": {
                    "description": "响应类型，该字段用于筛选，可配置S/F/E",
                    "type": "string",
                    "example": "S"
                },
                "zhCn": {
                    "description": "中文响应文言",
                    "type": "string",
                    "example": "ZhCn"
                }
            }
        },
        "platModel.ResponseNextValReq": {
            "type": "object",
            "required": [
                "serviceCode",
                "type"
            ],
            "properties": {
                "serviceCode": {
                    "description": "业务ID，来源于字典，指定响应码归属业务",
                    "type": "string",
                    "maxLength": 3,
                    "example": "0"
                },
                "type": {
                    "description": "响应类型，该字段用于筛选，可配置S/F/E",
                    "type": "string",
                    "enum": [
                        "S",
                        "F",
                        "E"
                    ],
                    "example": "S"
                }
            }
        },
        "platModel.ResponsePageReq": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应码，支持模糊查询",
                    "type": "string",
                    "example": "20"
                },
                "current": {
                    "description": "当前页",
                    "type": "integer",
                    "example": 1
                },
                "pageSize": {
                    "description": "单页数量",
                    "type": "integer",
                    "example": 10
                },
                "serviceCode": {
                    "description": "业务ID，来源于字典，指定响应码归属业务",
                    "type": "string",
                    "example": "0"
                },
                "type": {
                    "description": "响应类型，该字段用于筛选，可配置S/F/E",
                    "type": "string",
                    "example": "S"
                }
            }
        },
        "platModel.ResponsePageRes": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应码",
                    "type": "string",
                    "example": "20101"
                },
                "enUs": {
                    "description": "英文响应文言",
                    "type": "string",
                    "example": "EnUs"
                },
                "id": {
                    "description": "数据ID",
                    "type": "integer",
                    "example": 1
                },
                "mark": {
                    "description": "变更标识 0可变更 1禁止变更",
                    "type": "string",
                    "example": "1"
                },
                "remark": {
                    "description": "其他备注信息",
                    "type": "string",
                    "example": "Remark"
                },
                "serviceCode": {
                    "description": "业务ID，来源于字典，指定响应码归属业务",
                    "type": "string",
                    "example": "1"
                },
                "type": {
                    "description": "响应类型，该字段用于筛选，可配置S/F/E",
                    "type": "string",
                    "example": "S"
                },
                "zhCn": {
                    "description": "中文响应文言",
                    "type": "string",
                    "example": "ZhCn"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Token": {
            "type": "apiKey",
            "name": "Token",
            "in": "header"
        }
    },
...
}